third_party/WebKit/LayoutTests/bluetooth/descriptor/readValue/gen-descriptor-is-removed.html

<!-- Generated by //third_party/WebKit/LayoutTests/bluetooth/generate.py -->
<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
  let val = new Uint8Array([1]);
  return setBluetoothFakeAdapter('DisconnectingHealthThermometerAdapter')
      .then(
          () => requestDeviceWithKeyDown(
              {filters: [{services: ['health_thermometer']}]}))
      .then(device => device.gatt.connect())
      .then(gattServer => gattServer.getPrimaryService('health_thermometer'))
      .then(service => service.getCharacteristic('measurement_interval'))
      .then(
          characteristic => characteristic.getDescriptor(user_description.name))
      .then(descriptor => {
        return setBluetoothFakeAdapter(
                   'MissingDescriptorsDisconnectingHealthThermometerAdapter')
            .then(
                () => assert_promise_rejects_with_message(
                    descriptor.readValue(),
                    new DOMException(
                        'GATT Descriptor no longer exists.',
                        'InvalidStateError'),
                    'Descriptor got removed.'));
      });
}, 'Descriptor gets removed. Reject with InvalidStateError.');

</script>
